requirements.txt
dbus-python
paho-mqtt
pybluez
